requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
msoffcrypto-tool>=5.0.0
matplotlib>=3.7.0
